feat: Add mypy --strict compliance (Quality Scale strict-typing)
Add PEP-561 py.typed marker and full type annotations across all 47
source files to pass mypy --strict with zero errors. Replace Any with
concrete types via TYPE_CHECKING guards to resolve circular imports,
use Mapping for read-only MappingProxyType parameters, type Event
generics with EventStateChangedData, and add mixin attribute stubs
for TriggerConfigMixin. Replace assert guards with graceful aborts
and add defensive int() casts on config option divisions.

+++ custom_components/maintenance_supporter/helpers/interval_analyzer.py
@@ -504,7 +504,7 @@ class IntervalAnalyzer:
             return 0
         try:
             t = eta * ((-math.log(reliability)) ** (1 / beta))
-            return max(1, round(t))
+            return int(max(1, round(t)))
         except (ValueError, ZeroDivisionError, OverflowError):
             return 0
